+++ src/app/core/services/api/legacy-managent-api-client/api/api.ts
@@ -1,6 +1,5 @@
 // @ts-nocheck
 import {ApplicationObservabilityService} from './applicationObservability.service';
-import {AssetService} from './asset.service';
 import {CatalogService} from './catalog.service';
 import {ContractAgreementService} from './contractAgreement.service';
 import {ContractDefinitionService} from './contractDefinition.service';
@@ -9,12 +8,9 @@ import {DataplaneSelectorService} from './dataplaneSelector.service';
 import {DefaultService} from './default.service';
 import {HTTPProvisionerWebhookService} from './hTTPProvisionerWebhook.service';
 import {IdentityHubService} from './identityHub.service';
-import {PolicyService} from './policy.service';
 import {TransferProcessService} from './transferProcess.service';
 
 export * from './applicationObservability.service';
-
-export * from './asset.service';
 
 export * from './catalog.service';
 
@@ -32,13 +28,10 @@ export * from './hTTPProvisionerWebhook.service';
 
 export * from './identityHub.service';
 
-export * from './policy.service';
-
 export * from './transferProcess.service';
 
 export const APIS = [
   ApplicationObservabilityService,
-  AssetService,
   CatalogService,
   ContractAgreementService,
   ContractDefinitionService,
@@ -47,6 +40,5 @@ export const APIS = [
   DefaultService,
   HTTPProvisionerWebhookService,
   IdentityHubService,
-  PolicyService,
   TransferProcessService,
 ];