+++ src/app/core/services/api/policy-type-ext.ts
@@ -1,16 +1,15 @@
-import {Constraint} from './legacy-managent-api-client';
+import {UiPolicyConstraintOperatorEnum} from '@sovity.de/edc-client';
 
-export const ExpressionLeftSideConstants = {
+export const PolicyLeftExpressions = {
   PolicyEvaluationTime: 'POLICY_EVALUATION_TIME',
+  ReferringConnector: 'REFERRING_CONNECTOR',
 };
 
-export const EdcTypes = {
-  AtomicConstraint: 'AtomicConstraint',
-  LiteralExpression: 'dataspaceconnector:literalexpression',
-  Permission: 'dataspaceconnector:permission',
-};
+export type OperatorDto =
+  | UiPolicyConstraintOperatorEnum
+  | UiPolicyConstraintOperatorEnum;
 
-export const OperatorSymbols = {
+export const OPERATOR_SYMBOLS: Record<OperatorDto, string> = {
   GT: '>',
   LT: '<',
   EQ: '=',
@@ -18,18 +17,9 @@ export const OperatorSymbols = {
   GEQ: '≥',
   LEQ: '≤',
   IN: '∈',
+  HAS_PART: 'HAS PART',
+  IS_A: 'IS A',
+  IS_ALL_OF: 'IS ALL OF',
+  IS_ANY_OF: 'IS ANY OF',
+  IS_NONE_OF: 'IS NONE OF',
 };
-
-export type Operator = keyof typeof OperatorSymbols;
-
-export interface AtomicConstraint extends Constraint {
-  edctype: string;
-  leftExpression: unknown;
-  rightExpression: unknown;
-  operator: Operator;
-}
-
-export interface LiteralExpression {
-  edctype: string;
-  value: string;
-}