+++ src/app/routes/connector-ui/contract-definition-page/contract-definition-editor-dialog/contract-definition-editor-dialog-form.ts
@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
 import {FormBuilder, FormGroup, Validators} from '@angular/forms';
-import {PolicyDefinition} from '../../../../core/services/api/legacy-managent-api-client';
+import {PolicyDefinitionDto} from '@sovity.de/edc-client';
 import {Asset} from '../../../../core/services/models/asset';
 import {noWhitespaceValidator} from '../../../../core/validators/no-whitespace-validator';
 import {
@@ -27,8 +27,8 @@ export class ContractDefinitionEditorDialogForm {
   buildFormGroup(): FormGroup<ContractDefinitionEditorDialogFormModel> {
     return this.formBuilder.nonNullable.group({
       id: ['', [Validators.required, noWhitespaceValidator]],
-      accessPolicy: [null as PolicyDefinition | null, Validators.required],
-      contractPolicy: [null as PolicyDefinition | null, Validators.required],
+      accessPolicy: [null as PolicyDefinitionDto | null, Validators.required],
+      contractPolicy: [null as PolicyDefinitionDto | null, Validators.required],
       assets: [[] as Asset[], Validators.required],
     });
   }