+++ src/app/routes/connector-ui/policy-definition-page/policy-cards/policy-card.ts
@@ -1,12 +1,10 @@
-import {PolicyDefinition} from '../../../../core/services/api/legacy-managent-api-client';
-
 export interface PolicyCard {
   id: string;
   isRegular: boolean;
   irregularities: string[];
   constraints: PolicyCardConstraint[];
 
-  objectForJson: PolicyDefinition;
+  objectForJson: any;
 }
 
 export interface PolicyCardConstraint {